fix4

+++ gyros.py
@@ -61,8 +61,8 @@ if __name__ == "__main__":
             acclY_scaled = acc_y * .000061 * 9.80665
             acclZ_scaled = acc_z * .000061 * 9.80665
              
-            x_angle: float =+ round(get_x_rotation(acclX_scaled, acclY_scaled, acclZ_scaled),0)
-            y_angle: float=+ round(get_y_rotation(acclX_scaled, acclY_scaled, acclZ_scaled),0)
+            x_angle: int =+ round(get_x_rotation(acclX_scaled, acclY_scaled, acclZ_scaled),0)
+            y_angle: int=+ round(get_y_rotation(acclX_scaled, acclY_scaled, acclZ_scaled),0)
             count =+ 1
             
             #print("X rotation: ", round(x_angle, 2))
